+++ CheatSheetPlayer.cs
@@ -12,7 +12,7 @@ namespace CheatSheet
 		public Item[] ExtraAccessories = new Item[MaxExtraAccessories];
 		public int numberExtraAccessoriesEnabled = 0;
 
-		public override void UpdateEquips(ref bool wallSpeedBuff, ref bool tileSpeedBuff, ref bool tileRangeBuff)
+		public override void UpdateEquips()
 		{
 			for (int i = 0; i < numberExtraAccessoriesEnabled; i++)
 			{
@@ -20,7 +20,7 @@ namespace CheatSheet
 			}
 			for (int i = 0; i < numberExtraAccessoriesEnabled; i++)
 			{
-				player.VanillaUpdateAccessory(player.whoAmI, ExtraAccessories[i], false, ref wallSpeedBuff, ref tileSpeedBuff, ref tileRangeBuff);
+				player.ApplyEquipFunctional(ExtraAccessories[i], false);
 			}
 		}
